# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateRepoRequest": {
        "description": "Body for `POST /api/auth/github/repos/new`.",
        "properties": {
          "deploy_to_cloudflare_pages": {
            "default": false,
            "title": "Deploy To Cloudflare Pages",
            "type": "boolean"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "private": {
            "default": true,
            "title": "Private",
            "type": "boolean"
          },
          "register_in_pravi": {
            "default": true,
            "title": "Register In Pravi",
            "type": "boolean"
          },
          "template": {
            "default": "vite-react-static",
            "title": "Template",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateRepoRequest",
        "type": "object"
      },
      "CreateRepoResult": {
        "description": "Outcome of a new-repo create. Surfaces what happened on each leg\nso the UI can render success / partial-success / what-to-do-next.",
        "properties": {
          "initial_commit_pushed": {
            "title": "Initial Commit Pushed",
            "type": "boolean"
          },
          "pages": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PagesProjectOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "pages_skipped_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pages Skipped Reason"
          },
          "pravi_repo_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pravi Repo Id"
          },
          "repo": {
            "$ref": "#/components/schemas/GitHubRepoOut"
          }
        },
        "required": [
          "repo",
          "initial_commit_pushed"
        ],
        "title": "CreateRepoResult",
        "type": "object"
      },
...
      "GitHubIssueLabelOut": {
        "properties": {
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "GitHubIssueLabelOut",
        "type": "object"
      },
      "GitHubIssueOut": {
        "description": "One issue from the issues-list endpoint. Filters out pull requests.",
        "properties": {
          "body": {
            "default": "",
            "title": "Body",
            "type": "string"
          },
          "comments": {
            "default": 0,
            "title": "Comments",
            "type": "integer"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "html_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Html Url"
          },
          "labels": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/GitHubIssueLabelOut"
            },
            "title": "Labels",
            "type": "array"
          },
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "state": {
            "default": "open",
            "title": "State",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "user_avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Avatar Url"
          },
          "user_login": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Login"
          }
        },
        "required": [
          "number",
          "title"
        ],
        "title": "GitHubIssueOut",
        "type": "object"
      },
...
      "GitHubRepoOut": {
        "description": "One repo from `/api/auth/github/repos/search`.\n\nMirrors what the new-ticket picker needs: enough to display + enough to\nlater clone. `clone_url` is the HTTPS form; a token is injected at clone\ntime, never stored at rest on the repo's git remote.",
        "properties": {
          "clone_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clone Url"
          },
          "default_branch": {
            "default": "main",
            "title": "Default Branch",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "open_issues_count": {
            "default": 0,
            "title": "Open Issues Count",
            "type": "integer"
          },
          "owner": {
            "title": "Owner",
            "type": "string"
          },
          "private": {
            "default": false,
            "title": "Private",
            "type": "boolean"
          },
          "ssh_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssh Url"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "owner",
          "name",
          "full_name"
        ],
        "title": "GitHubRepoOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PagesProjectOut": {
        "properties": {
          "canonical_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Url"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pages_url": {
            "title": "Pages Url",
            "type": "string"
          },
          "subdomain": {
            "title": "Subdomain",
            "type": "string"
          }
        },
        "required": [
          "name",
          "subdomain",
          "pages_url"
        ],
        "title": "PagesProjectOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "pravi",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/github/repos/new": {
      "post": {
        "description": "Create a brand-new GitHub repo, seed it with a starter template,\noptionally wire it up to Cloudflare Pages for auto-deploy, and\nregister it as a pravi Repo so it's immediately usable as a ticket\ntarget.\n\nSteps:\n  1. POST /user/repos \u2192 empty repo on GitHub\n  2. Local temp checkout, write template files, push initial commit\n  3. (Optional) Cloudflare Pages project tied to this repo\n  4. (Optional) Register a pravi Repo row pointing at the local clone",
        "operationId": "create_new_repo_api_auth_github_repos_new_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRepoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateRepoResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create New Repo"
      }
    },
    "/api/auth/github/repos/search": {
      "get": {
        "description": "Search GitHub repos accessible to the active OAuth connection.\n\nUsed by the new-ticket form to let the user pick a repo without typing\na local path. Empty `q` returns the most-recently-pushed repos.",
        "operationId": "search_repos_api_auth_github_repos_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GitHubRepoOut"
                  },
                  "title": "Response Search Repos Api Auth Github Repos Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Repos"
      }
    },
    "/api/auth/github/repos/{owner}/{name}/issues": {
      "get": {
        "description": "List issues on a connected GitHub repo. PRs are filtered out.\n\nUsed by the /issues page to scan + import as pravi tickets. `state` is\n\"open\" | \"closed\" | \"all\"; `labels` is comma-separated.",
        "operationId": "list_repo_issues_api_auth_github_repos__owner___name__issues_get",
        "parameters": [
          {
            "in": "path",
            "name": "owner",
            "required": true,
            "schema": {
              "title": "Owner",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "default": "open",
              "title": "State",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "labels",
            "required": false,
            "schema": {
              "default": "",
              "title": "Labels",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GitHubIssueOut"
                  },
                  "title": "Response List Repo Issues Api Auth Github Repos  Owner   Name  Issues Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Repo Issues"
      }
    },
...
  }
}